:recycle: separate Cell component in ColumnTable, RowTable.
